create api call for login, dashboard and platform

+++ TechnicalTest/Models/ChartDashboardDto.cs
@@ -3,5 +3,5 @@ namespace TechnicalTest.Models;
 public class ChartDashboardDto
 {
     public string name { get; set; } = "";
-    public long value { get; set; }
+    public int value { get; set; }
 }
